Use long type for PunchCard statistics in case just in case there are some really big counts

+++ Octokit/Models/Response/PunchCard.cs
@@ -12,7 +12,7 @@ namespace Octokit
     {
         public PunchCard() { }
 
-        public PunchCard(IEnumerable<IList<int>> punchCardData)
+        public PunchCard(IEnumerable<IList<long>> punchCardData)
         {
             Ensure.ArgumentNotNull(punchCardData, nameof(punchCardData));
             PunchPoints = new ReadOnlyCollection<PunchCardPoint>(
@@ -36,7 +36,7 @@ namespace Octokit
         /// <param name="dayOfWeek">The day of the week to query</param>
         /// <param name="hourOfDay">The hour in 24 hour time. 0-23.</param>
         /// <returns>The total number of commits made.</returns>
-        public int GetCommitCountFor(DayOfWeek dayOfWeek, int hourOfDay)
+        public long GetCommitCountFor(DayOfWeek dayOfWeek, int hourOfDay)
         {
             var punchPoint = PunchPoints.SingleOrDefault(point => point.DayOfWeek == dayOfWeek && point.HourOfTheDay == hourOfDay);
             return punchPoint == null ? 0 : punchPoint.CommitCount;
